+++ essentials-base/src/main/kotlin/io/github/nyliummc/essentials/entities/EssentialsDatabase.kt
@@ -35,14 +35,20 @@ import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ConcurrentLinkedQueue
 import kotlin.concurrent.thread
 import io.github.nyliummc.essentials.api.EssentialsDatabase as APIEssentialsDatabase
+
 object EssentialsDatabase : APIEssentialsDatabase {
-    var running = true
-    var queue = ConcurrentLinkedQueue<Pair<Transaction.() -> Any, CompletableFuture<Any>>>()
-    private val databaseThread = thread(start = true, name = "Essentials Database Thread") {
+    @Volatile private var running = true
+    private val queue = ConcurrentLinkedQueue<Pair<Transaction.() -> Any, CompletableFuture<Any>>>()
+    private val databaseThread = thread(start=true, isDaemon=false, name="Essentials Database Thread") {
         while (running) {
-            val pair = queue.poll() ?: continue
+            val pair = queue.poll()
+            if (pair == null) {
+                Thread.sleep(20)  // 20ms to not lag thread
+                continue
+            }
             val value = dbTransaction {  // Because recursion
-                pair.first.invoke(this)
+                val x = pair.first.invoke(this)
+                x
             }
             pair.second.complete(value)
         }
@@ -55,6 +61,8 @@ object EssentialsDatabase : APIEssentialsDatabase {
     private val databaseName = "essentials"
 
     fun disconnect() {
+        running = false
+        databaseThread.join()
         try {
             TransactionManager.closeAndUnregister(db)
         } catch (e: UninitializedPropertyAccessException) {
@@ -104,6 +112,7 @@ object EssentialsDatabase : APIEssentialsDatabase {
 
     override fun <T> transaction(callback: Transaction.() -> T): CompletableFuture<T> {
         val fut = CompletableFuture<T>()
+        println("Waiting for future ${fut.hashCode()}")
         queue.add(Pair(callback as Transaction.() -> Any, fut as CompletableFuture<Any>))
         return fut
     }